refactor: move referrals related schemas into separate file
